move sort function to existing sort-by.ts file
