+++ src/utilities/sort-by.ts
@@ -37,3 +37,15 @@ export const sortByNumber = (order: SortOrder, propertyToCompare: PropertyToComp
     }
     return -result; // ascending order
   });
+
+// Sort function for Array.sort comparator functions
+export const sortAlphabetically = (a: string, b: string): number => {
+  const aLower = a.toLowerCase();
+  const bLower = b.toLowerCase();
+
+  if (aLower < bLower) { return -1; }
+  if (aLower > bLower) { return 1; }
+  return 0;
+}
+
+export default sortAlphabetically;
